sky: Enable sysfs fallback
As documented by the Kconfig help text, the right way to enable the
fallback (if needed) is to echo 1 into
/proc/sys/kernel/firmware_config/force_sysfs_fallback

Setting the fallback globally causes delays when using the in-kernel
firmware loader if the driver module is loaded in first stage init, even
if the driver would correctly retry direct firmware loading some time
later.

Any driver relying on the FW_LOADER_USER_HELPER should be migrated to
use the kernel's direct firmware loader so that init/ueventd does not
need to load the firmware for the driver, which is less secure.